feat: Add a System construct to represent the underlying system.
This System struct forms the starting step to list all the pools in the
underlying system.

The unexported cmd struct also acts as a means to invoke all the
commands in this system.

Pool, FileSystem, Snapshot all provide a cmd() helper function to
retrieve this command handle for invoking any zpool/zfs commands.

The SystemConfig input for creating a new System object provides a few
unexported fields to inject a custom command invoker. This will be used
in the future to allow using a fake zpool in tests.